+++ wn/_exceptions.py
@@ -17,6 +17,11 @@ class ConfigurationError(Error):
     __module__ = 'wn'
 
 
+class ProjectError(Error):
+    """Raised when a project is not found or on errors defined in the index."""
+    __module__ = 'wn'
+
+
 class WnWarning(Warning):
     """Generic warning class for dubious worndet operations."""